requirements.txt
asn1crypto==0.23.0
bcrypt==3.1.4
cffi==1.11.2
click==6.7
cryptography==2.1.2
enum34==1.1.6
Fabric3==1.13.1.post1
gitdb2==2.0.3
GitPython==2.1.7
idna==2.6
ipaddress==1.0.18
paramiko==2.3.1
pyaml==17.10.0
pyasn1==0.3.7
pycparser==2.18
PyNaCl==1.1.2
PyYAML==3.12
six==1.11.0
smmap2==2.0.3
